Adds test type FK to bookings
Adds a foreign key column 'TestTypeId' to the 'Bookings' table, linking it to the 'TestTypes' table.
This change allows tracking the type of test associated with each booking.
Updates the database model and migrations to reflect this new relationship.
Implements a first-time booking check to not allow the user to book multiple times the same test for the same application,
also implements the retake test logic for booking the same test.
Refactor the create local driving license application to use the orchestrator pattern to separate the services calls

Add Documentation to some code.