Renames and fixes application entity
Renames the 'Applicataions' table to 'Applications' to adhere to naming conventions.
Updates the Application entity configuration to specify the table name and correct a typo in the DbSet property name in the DbContext.
Adds foreign key relationships to the applications table.
Updates the retake test application to be linked with the applicaiton table.

Fixes issues related to table naming and entity relationships