Corrects table name typo and adds new features
Fixes a typo in the table name "Applicataions" to "Applications" across migrations, foreign key constraints, and indexes.
Adds .vs directory to .gitignore to ignore visual studio specific files.
Adds a foreign key constraint for service purposes and service categories to the applications table, and also adds a new table for RetakeTestApplication.
Adds an event to the test creator service so that the application can subscribe for test creation.
Also, updates the service fees DTO and service to align the service purpose field name across the app