Adds entity configurations
Adds configurations for LDLApplicationsAllowedToRetakeATest, Employee, and Admin entities.
Adds dummy data for Employee and Admin for testing purposes.
Also, fixes a column name in LDLApplicationsAllowedToRetakeATest table in migrations.
Updates the GetLDLAppsAllowedToRetakATest to return the columns and fixes the type of the returned object.
Updates the test creation service to return testDTO.
Also modifies the testRetakeApplicationCreator service, to fetch the test record and use it for its operations.

There an issue with DBcontext with Dependency Injection and the table-value functions