Adds retake test application event
Implements event for when a retake test application is created,
allowing subscribers to react to the creation of a retake test,
for example, updating LDL status.
Also, fixing the validation of retake tests to consider TestTypeId.
Moves RetakeTestApplicationRetriever to its own file.
Removes some unused code in the appointment update service
and makes it event-based as well.